+++ test/src/log-transport-test.js
@@ -9,6 +9,8 @@ import { FakeApplicationInsights } from '../../src/fake-applicationinsights.js';
 
 const filePath = fileURLToPath(import.meta.url);
 
+const tagKeys = new Contracts.ContextTagKeys();
+
 describe('log transport', () => {
   const connectionString = `InstrumentationKey=${randomUUID()};IngestionEndpoint=https://ingestion.local;LiveEndpoint=https://livemonitor.local/`;
 
@@ -21,7 +23,7 @@ describe('log transport', () => {
       config: { maxBatchSize: 1, disableStatsbeat: true },
     });
 
-    logger = pino({ level: 'trace' }, transport);
+    logger = pino({ level: 'trace', mixin }, transport);
   });
   after(() => {
     transport.destroy();
@@ -37,6 +39,8 @@ describe('log transport', () => {
     const msg = await expectMessage;
 
     expect(msg.body.data.baseData).to.deep.include({ severityLevel: Contracts.SeverityLevel.Verbose, message: 'foo' });
+
+    expect(msg.body.tags).to.deep.include({ [tagKeys.userId]: 'uzer' });
   });
 
   it('logs info', async () => {
@@ -47,6 +51,16 @@ describe('log transport', () => {
     const msg = await expectMessage;
 
     expect(msg.body.data.baseData).to.deep.include({ severityLevel: Contracts.SeverityLevel.Information, message: 'foo' });
+  });
+
+  it('logs info with tag overrides', async () => {
+    const expectMessage = fakeAI.expectMessageData();
+
+    logger.info({ bar: 'baz', tagOverrides: { [tagKeys.userAuthUserId]: 'Jan Bananberg' } }, 'foo');
+
+    const msg = await expectMessage;
+
+    expect(msg.body.tags).to.deep.include({ [tagKeys.userAuthUserId]: 'Jan Bananberg' });
   });
 
   it('logs warn', async () => {
@@ -107,6 +121,21 @@ describe('log transport', () => {
     expect(exception.parsedStack[0].fileName, 'stack file name').to.include(filePath);
   });
 
+  it('logs exception with tag overrides', async () => {
+    const expectMessage = fakeAI.expectMessageData();
+    const expectException = fakeAI.expectExceptionData();
+
+    logger.error(new Error('bar'), 'foo');
+
+    const msg = await expectMessage;
+
+    expect(msg.body.tags).to.deep.include({ [tagKeys.userId]: 'uzer' });
+
+    const err = await expectException;
+
+    expect(err.body.tags).to.deep.include({ [tagKeys.userId]: 'uzer' });
+  });
+
   it('logs fatal', async () => {
     const expectMessage = fakeAI.expectMessageData();
     const expectException = fakeAI.expectExceptionData();
@@ -147,3 +176,9 @@ describe('log transport', () => {
     expect(msg.body.time).to.equal(new Date().toISOString());
   });
 });
+
+function mixin(context) {
+  return {
+    tagOverrides: { [tagKeys.userId]: 'uzer', ...context.tagOverrides },
+  };
+}